setting default comment days to unlimited.  this is mainly for folks using metaweblogapi, etc, who can't set this value and always end up with only 7 days of comments :/
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/roller/trunk@225538 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/roller/pojos/WeblogEntryData.java
@@ -57,7 +57,7 @@ public class WeblogEntryData extends org.roller.pojos.PersistentObject
     protected WebsiteData mWebsite=null;
     protected String mPlugins;
     protected Boolean allowComments = Boolean.TRUE;
-    protected Integer commentDays = new Integer(7);
+    protected Integer commentDays = new Integer(0);
     protected Boolean rightToLeft = Boolean.FALSE;
     protected Boolean pinnedToMain = Boolean.FALSE;
